+++ src/shared/context/UiContextProvider.tsx
@@ -1,6 +1,7 @@
 'use client'
 import React, { useEffect, useState } from 'react'
 
+import updateHashState from '@/shared/helpers/updateHashState'
 import useUiContextFromHashParams from '@/shared/hooks/useUiContextFromHashParams'
 
 import type { UiContextType } from './uiContext'
@@ -14,9 +15,7 @@ const UiProvider = ({ children }: { children: React.ReactNode }) => {
   })
 
   useEffect(() => {
-    window.location.hash = new URLSearchParams(
-      uiState as unknown as Record<string, string>,
-    ).toString()
+    updateHashState(uiState)
   }, [uiState])
 
   const updateUiState = (newState: Partial<UiContextType['uiState']>) => {
